+++ plugins/discourse-ai/spec/lib/agents/tool_runner_spec.rb
@@ -17,17 +17,9 @@ RSpec.describe DiscourseAi::Agents::ToolRunner do
   end
   fab!(:llm_model)
   fab!(:ai_secret)
-  fab!(:tag1) { Fabricate(:tag, name: "tag1") }
-  fab!(:tag2) { Fabricate(:tag, name: "tag2") }
-  fab!(:category) { Fabricate(:category, name: "Test Category", slug: "test-category") }
-  fab!(:topic) { Fabricate(:topic, category: category) }
-  fab!(:pm_topic, :private_message_topic)
   let(:llm) { DiscourseAi::Completions::Llm.proxy(llm_model.id) }
 
-  before do
-    enable_current_plugin
-    SiteSetting.tagging_enabled = true
-  end
+  before { enable_current_plugin }
 
   describe "#invoke" do
     it "can execute a simple script" do
@@ -95,723 +87,6 @@ RSpec.describe DiscourseAi::Agents::ToolRunner do
 
       expect(result["key"]).to eq(ai_secret.secret)
     end
-
-    it "can set tags on a topic" do
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editTopic(params.topic_id, { tags: ['tag1', 'tag2'] }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["tags"]).to contain_exactly("tag1", "tag2")
-      expect(topic.reload.tags.pluck(:name)).to contain_exactly("tag1", "tag2")
-    end
-
-    it "can get a topic with tags and first_post_id" do
-      Fabricate(:post, topic: topic)
-      tag = Fabricate(:tag, name: "test_tag")
-      topic.tags << tag
-      tool.update!(
-        script:
-          "function invoke(params) { const t = discourse.getTopic(params.topic_id); return { tags: t.tags, first_post_id: t.first_post_id }; }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["tags"]).to contain_exactly("test_tag")
-      expect(result["first_post_id"]).to eq(topic.first_post.id)
-    end
-
-    it "can append tags on a topic" do
-      old_tag = Fabricate(:tag, name: "old_tag")
-      topic.tags << old_tag
-      Fabricate(:tag, name: "new_tag")
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editTopic(params.topic_id, { tags: ['new_tag'] }, { append: true }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(topic.reload.tags.pluck(:name)).to contain_exactly("old_tag", "new_tag")
-    end
-
-    it "can set tags as a specific user" do
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editTopic(params.topic_id, { tags: ['tag1'] }, { username: 'system' }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["tags"]).to eq(["tag1"])
-      expect(topic.reload.tags.pluck(:name)).to contain_exactly("tag1")
-    end
-
-    it "publishes to MessageBus when setting tags" do
-      topic = Fabricate(:topic)
-      Fabricate(:post, topic: topic)
-      Fabricate(:tag, name: "msgbus_tag")
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.setTags(params.topic_id, ['msgbus_tag']); }",
-      )
-
-      messages =
-        MessageBus.track_publish("/topic/#{topic.id}") do
-          runner =
-            described_class.new(
-              parameters: {
-                topic_id: topic.id,
-              },
-              llm: llm,
-              bot_user: bot_user,
-              tool: tool,
-            )
-          result = runner.invoke
-          expect(result["success"]).to eq(true)
-        end
-
-      expect(messages).not_to be_empty
-      expect(messages.first.data[:type]).to eq(:revised)
-    end
-
-    it "can edit a post" do
-      post = Fabricate(:post)
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editPost(params.post_id, 'new raw content', { edit_reason: 'AI edit' }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            post_id: post.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(post.reload.raw).to eq("new raw content")
-      expect(post.edit_reason).to eq("AI edit")
-      expect(post.last_editor_id).to eq(bot_user.id)
-    end
-
-    it "can edit a post as a specific user" do
-      other_user = Fabricate(:user, admin: true)
-      post = Fabricate(:post)
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editPost(params.post_id, 'new raw content', { username: '#{other_user.username}' }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            post_id: post.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(post.reload.last_editor_id).to eq(other_user.id)
-    end
-
-    it "denies editing a post when the user lacks permission" do
-      post_author = Fabricate(:user)
-      unprivileged_user = Fabricate(:user)
-      post = Fabricate(:post, user: post_author)
-      original_raw = post.raw
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.editPost(params.post_id, 'hacked content', { username: '#{unprivileged_user.username}' }); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            post_id: post.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      expect { runner.invoke }.to raise_error(/Permission denied/)
-      expect(post.reload.raw).to eq(original_raw)
-    end
-
-    it "can generate JSON from LLM" do
-      tool.update!(script: "function invoke() { return llm.generate('test', { json: true }); }")
-
-      DiscourseAi::Completions::Llm.with_prepared_responses(
-        ['{"key": "value"}'],
-      ) do |_, _, _, prompt_options|
-        runner = described_class.new(parameters: {}, llm: llm, bot_user: bot_user, tool: tool)
-        result = runner.invoke
-        expect(result).to eq({ "key" => "value" })
-
-        expect(prompt_options.last[:response_format]).to eq({ "type" => "json_object" })
-      end
-    end
-
-    it "can get a topic with category info" do
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            const t = discourse.getTopic(params.topic_id);
-            return {
-              category_id: t.category_id,
-              category_name: t.category_name,
-              category_slug: t.category_slug
-            };
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["category_id"]).to eq(category.id)
-      expect(result["category_name"]).to eq("Test Category")
-      expect(result["category_slug"]).to eq("test-category")
-    end
-
-    it "can set category on a topic by slug" do
-      new_category = Fabricate(:category, slug: "new-category")
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, { category: "new-category" });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["category_id"]).to eq(new_category.id)
-      expect(topic.reload.category_id).to eq(new_category.id)
-    end
-
-    it "can set category on a topic by ID" do
-      new_category = Fabricate(:category)
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, { category: params.category_id });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-            category_id: new_category.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["category_id"]).to eq(new_category.id)
-      expect(topic.reload.category_id).to eq(new_category.id)
-    end
-
-    it "returns error when setting category on private message" do
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            try {
-              return discourse.editTopic(params.topic_id, { category: params.category_id });
-            } catch(e) {
-              return { error: e.message };
-            }
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: pm_topic.id,
-            category_id: category.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["error"]).to include("private")
-    end
-
-    it "can unlist a topic" do
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, { visible: false });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["visible"]).to eq(false)
-      expect(topic.reload.visible).to eq(false)
-      expect(topic.visibility_reason_id).to eq(Topic.visibility_reasons[:manually_unlisted])
-    end
-
-    it "can relist a topic" do
-      topic.update!(visible: false)
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, { visible: true });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["visible"]).to eq(true)
-      expect(topic.reload.visible).to eq(true)
-      expect(topic.visibility_reason_id).to eq(Topic.visibility_reasons[:manually_relisted])
-    end
-
-    it "editTopic throws error for non-existent topic" do
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            try {
-              return discourse.editTopic(999999, { tags: ['tag1'] });
-            } catch(e) {
-              return { thrown: true, message: e.message };
-            }
-          }
-        JS
-      runner = described_class.new(parameters: {}, llm: llm, bot_user: bot_user, tool: tool)
-      result = runner.invoke
-      expect(result["thrown"]).to eq(true)
-      expect(result["message"]).to include("not found")
-    end
-
-    it "can edit multiple topic properties at once" do
-      new_category = Fabricate(:category, slug: "multi-edit")
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, {
-              category: "multi-edit",
-              tags: ['tag1', 'tag2'],
-              visible: false
-            });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["category_id"]).to eq(new_category.id)
-      expect(result["topic"]["tags"]).to contain_exactly("tag1", "tag2")
-      expect(result["topic"]["visible"]).to eq(false)
-      topic.reload
-      expect(topic.category_id).to eq(new_category.id)
-      expect(topic.tags.pluck(:name)).to contain_exactly("tag1", "tag2")
-      expect(topic.visible).to eq(false)
-    end
-
-    it "can set category on a topic by name" do
-      new_category = Fabricate(:category, name: "New Category Name", slug: "different-slug")
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            return discourse.editTopic(params.topic_id, { category: "New Category Name" });
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(result["topic"]["category_id"]).to eq(new_category.id)
-      expect(topic.reload.category_id).to eq(new_category.id)
-    end
-
-    it "returns error when user cannot move topic to category" do
-      regular_user = Fabricate(:user)
-      new_category = Fabricate(:category)
-      new_category.set_permissions(staff: :full)
-      new_category.save!
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            try {
-              return discourse.editTopic(params.topic_id, { category: params.category_id }, { username: params.username });
-            } catch(e) {
-              return { error: e.message };
-            }
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-            category_id: new_category.id,
-            username: regular_user.username,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["error"]).to include("Permission denied")
-    end
-
-    it "returns error when user cannot toggle topic visibility" do
-      regular_user = Fabricate(:user)
-      tool.update!(script: <<~JS)
-          function invoke(params) {
-            try {
-              return discourse.editTopic(params.topic_id, { visible: false }, { username: params.username });
-            } catch(e) {
-              return { error: e.message };
-            }
-          }
-        JS
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-            username: regular_user.username,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["error"]).to include("Permission denied")
-    end
-
-    it "setTags alias works for backwards compatibility" do
-      tool.update!(
-        script:
-          "function invoke(params) { return discourse.setTags(params.topic_id, ['tag1', 'tag2']); }",
-      )
-      runner =
-        described_class.new(
-          parameters: {
-            topic_id: topic.id,
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-      expect(result["success"]).to eq(true)
-      expect(topic.reload.tags.pluck(:name)).to contain_exactly("tag1", "tag2")
-    end
-
-    describe "guardian permission checks" do
-      fab!(:private_category) do
-        Fabricate(:category).tap do |c|
-          c.set_permissions(staff: :full)
-          c.save!
-        end
-      end
-      fab!(:private_topic) { Fabricate(:topic, category: private_category) }
-
-      def run_tool(script, parameters)
-        tool.update!(script: script)
-        described_class.new(parameters: parameters, llm: llm, bot_user: bot_user, tool: tool).invoke
-      end
-
-      it "denies regular user access to restricted topics and categories" do
-        params = { username: user.username }
-
-        expect {
-          run_tool(
-            "function invoke(params) { return discourse.editTopic(params.topic_id, { visible: false }, { username: params.username }); }",
-            params.merge(topic_id: private_topic.id),
-          )
-        }.to raise_error(MiniRacer::RuntimeError, /Permission denied/)
-
-        expect {
-          run_tool(
-            'function invoke(params) { return discourse.createTopic({ title: "Test title", raw: "Test body", category_id: params.category_id, username: params.username }); }',
-            params.merge(category_id: private_category.id),
-          )
-        }.to raise_error(MiniRacer::RuntimeError, /Permission denied/)
-
-        expect {
-          run_tool(
-            'function invoke(params) { return discourse.createPost({ topic_id: params.topic_id, raw: "Test reply", username: params.username }); }',
-            params.merge(topic_id: private_topic.id),
-          )
-        }.to raise_error(MiniRacer::RuntimeError, /Permission denied/)
-      end
-
-      it "defaults to bot_user when no username is provided" do
-        new_category = Fabricate(:category)
-        result =
-          run_tool(
-            'function invoke(params) { return discourse.createTopic({ title: "Bot topic title here", raw: "Bot topic body content", category_id: params.category_id }); }',
-            { category_id: new_category.id },
-          )
-        expect(result["success"]).to eq(true)
-        expect(Topic.find(result["topic_id"]).user_id).to eq(bot_user.id)
-
-        result =
-          run_tool(
-            'function invoke(params) { return discourse.createPost({ topic_id: params.topic_id, raw: "Bot reply content" }); }',
-            { topic_id: topic.id },
-          )
-        expect(result["success"]).to eq(true)
-        expect(Post.find(result["post_id"]).user_id).to eq(bot_user.id)
-
-        result =
-          run_tool(
-            "function invoke(params) { return discourse.editTopic(params.topic_id, { visible: false }); }",
-            { topic_id: topic.id },
-          )
-        expect(result["success"]).to eq(true)
-        expect(topic.reload.visible).to eq(false)
-      end
-
-      it "allows moderator to toggle visibility on a topic" do
-        mod = Fabricate(:moderator)
-        result =
-          run_tool(
-            "function invoke(params) { return discourse.editTopic(params.topic_id, { visible: false }, { username: params.username }); }",
-            { topic_id: topic.id, username: mod.username },
-          )
-        expect(result["success"]).to eq(true)
-        expect(topic.reload.visible).to eq(false)
-      end
-
-      it "allows user to createTopic in a permitted category" do
-        result =
-          run_tool(
-            'function invoke(params) { return discourse.createTopic({ title: "A valid topic title here", raw: "Some body content for the topic", category_id: params.category_id, username: params.username }); }',
-            { category_id: category.id, username: user.username },
-          )
-        expect(result["success"]).to eq(true)
-        expect(result["topic_id"]).to be_present
-      end
-
-      it "allows user to createPost in a permitted topic" do
-        result =
-          run_tool(
-            'function invoke(params) { return discourse.createPost({ topic_id: params.topic_id, raw: "A reply to the topic", username: params.username }); }',
-            { topic_id: topic.id, username: user.username },
-          )
-        expect(result["success"]).to eq(true)
-        expect(result["post_id"]).to be_present
-      end
-    end
-
-    describe "custom fields" do
-      it "can get a custom field from a post" do
-        post = Fabricate(:post)
-        post.custom_fields["test_key"] = "test_value"
-        post.save_custom_fields
-
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.getCustomField('post', params.post_id, 'test_key'); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              post_id: post.id,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        result = runner.invoke
-        expect(result).to eq("test_value")
-      end
-
-      it "can set a custom field on a post" do
-        post = Fabricate(:post)
-
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.setCustomField('post', params.post_id, 'ai_processed', 'yes'); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              post_id: post.id,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        result = runner.invoke
-
-        expect(result["success"]).to eq(true)
-        expect(post.reload.custom_fields["ai_processed"]).to eq("yes")
-      end
-
-      it "returns null for non-existent custom field" do
-        post = Fabricate(:post)
-
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.getCustomField('post', params.post_id, 'nonexistent'); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              post_id: post.id,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        result = runner.invoke
-        expect(result).to be_nil
-      end
-
-      it "works with topic custom fields" do
-        topic = Fabricate(:topic)
-
-        tool.update!(
-          script:
-            "function invoke(params) { discourse.setCustomField('topic', params.topic_id, 'processed', 'true'); return discourse.getCustomField('topic', params.topic_id, 'processed'); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              topic_id: topic.id,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        result = runner.invoke
-        expect(result).to eq("true")
-      end
-
-      it "throws error for invalid type in setCustomField" do
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.setCustomField('invalid', 1, 'key', 'val'); }",
-        )
-        runner = described_class.new(parameters: {}, llm: llm, bot_user: bot_user, tool: tool)
-        expect { runner.invoke }.to raise_error(MiniRacer::RuntimeError, /Invalid type/)
-      end
-
-      it "throws error for invalid type in getCustomField" do
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.getCustomField('invalid', 1, 'key'); }",
-        )
-        runner = described_class.new(parameters: {}, llm: llm, bot_user: bot_user, tool: tool)
-        expect { runner.invoke }.to raise_error(MiniRacer::RuntimeError, /Invalid type/)
-      end
-
-      it "throws error for key too long" do
-        post = Fabricate(:post)
-        long_key = "k" * (described_class::MAX_CUSTOM_FIELD_KEY_LENGTH + 1)
-
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.setCustomField('post', params.post_id, params.key, 'val'); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              post_id: post.id,
-              key: long_key,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        expect { runner.invoke }.to raise_error(MiniRacer::RuntimeError, /Key too long/)
-      end
-
-      it "throws error for value too long" do
-        post = Fabricate(:post)
-        long_value = "v" * (described_class::MAX_CUSTOM_FIELD_VALUE_LENGTH + 1)
-
-        tool.update!(
-          script:
-            "function invoke(params) { return discourse.setCustomField('post', params.post_id, 'key', params.value); }",
-        )
-        runner =
-          described_class.new(
-            parameters: {
-              post_id: post.id,
-              value: long_value,
-            },
-            llm: llm,
-            bot_user: bot_user,
-            tool: tool,
-          )
-        expect { runner.invoke }.to raise_error(MiniRacer::RuntimeError, /Value too long/)
-      end
-    end
   end
 
   describe "#has_custom_system_message?" do
@@ -856,107 +131,6 @@ RSpec.describe DiscourseAi::Agents::ToolRunner do
         JS
       runner = described_class.new(parameters: {}, llm: llm, bot_user: bot_user, tool: tool)
       expect(runner.custom_system_message).to be_nil
-    end
-  end
-
-  describe "#rag_get_file" do
-    before { SiteSetting.authorized_extensions = "md|txt" }
-
-    it "returns full file content in fragment order" do
-      upload = Fabricate(:upload, original_filename: "skill.md")
-      UploadReference.create!(upload: upload, target: tool)
-
-      RagDocumentFragment.create!(
-        target: tool,
-        upload: upload,
-        fragment: "Part 2",
-        fragment_number: 2,
-      )
-      RagDocumentFragment.create!(
-        target: tool,
-        upload: upload,
-        fragment: "Part 1",
-        fragment_number: 1,
-      )
-      RagDocumentFragment.create!(
-        target: tool,
-        upload: upload,
-        fragment: "Part 3",
-        fragment_number: 3,
-      )
-
-      tool.update!(
-        script: "function invoke(params) { return { content: index.getFile(params.filename) }; }",
-      )
-
-      runner =
-        described_class.new(
-          parameters: {
-            "filename" => "skill.md",
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-
-      expect(result["content"]).to eq("Part 1\nPart 2\nPart 3")
-    end
-
-    it "returns null for non-existent file" do
-      tool.update!(
-        script: "function invoke(params) { return { content: index.getFile(params.filename) }; }",
-      )
-
-      runner =
-        described_class.new(
-          parameters: {
-            "filename" => "nonexistent.md",
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-
-      expect(result["content"]).to be_nil
-    end
-
-    it "picks the latest upload when duplicate filenames exist" do
-      old_upload = Fabricate(:upload, original_filename: "skill.md")
-      new_upload = Fabricate(:upload, original_filename: "skill.md")
-      UploadReference.create!(upload: old_upload, target: tool)
-      UploadReference.create!(upload: new_upload, target: tool)
-
-      RagDocumentFragment.create!(
-        target: tool,
-        upload: old_upload,
-        fragment: "old content",
-        fragment_number: 1,
-      )
-      RagDocumentFragment.create!(
-        target: tool,
-        upload: new_upload,
-        fragment: "new content",
-        fragment_number: 1,
-      )
-
-      tool.update!(
-        script: "function invoke(params) { return { content: index.getFile(params.filename) }; }",
-      )
-
-      runner =
-        described_class.new(
-          parameters: {
-            "filename" => "skill.md",
-          },
-          llm: llm,
-          bot_user: bot_user,
-          tool: tool,
-        )
-      result = runner.invoke
-
-      expect(result["content"]).to eq("new content")
     end
   end
 end
